EDM-1685: Move tests to integration
