+++ test/integration/resources/resources_cli.go
